api만 밖으로 빼도록 변경
