+++ docs/exec-plans/tech-debt-tracker.md
@@ -12,5 +12,5 @@
 | ID | 영역 | 내용 | 영향도 | 상태 | 담당 | 생성일 | 목표일 |
 | --- | --- | --- | --- | --- | --- | --- | --- |
 | TD-001 | Docs | 제품/운영 지침 문서의 교차 링크 확장 필요 | 중 | open | team | 2026-04-24 | 2026-05-08 |
-| TD-002 | Tests | 아키텍처 경계 규칙(ArchUnit) 범위 확대 필요 | 중 | done | team | 2026-04-24 | 2026-04-27 |
+| TD-002 | Tests | 하이브리드 구조 기준 ArchUnit 경계 규칙 1차 정리 완료 | 중 | done | team | 2026-04-24 | 2026-04-27 |
 | TD-003 | CI | doc-gardening 리포트 기반 자동 PR 정책 정교화 필요 | 하 | open | team | 2026-04-24 | 2026-05-22 |
