+++ src/main/java/com/flodiback/api/meeting/MeetingController.java
@@ -4,11 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.flodiback.api.meeting.dto.CreateMeetingRequest;
-import com.flodiback.api.meeting.dto.CreateMeetingResponse;
-import com.flodiback.api.meeting.dto.MeetingDetailResponse;
-import com.flodiback.application.meeting.MeetingService;
-import com.flodiback.application.meeting.result.MeetingResult;
+import com.flodiback.domain.meeting.meeting.dto.CreateMeetingRequest;
+import com.flodiback.domain.meeting.meeting.dto.CreateMeetingResponse;
+import com.flodiback.domain.meeting.meeting.dto.MeetingDetailResponse;
+import com.flodiback.domain.meeting.meeting.service.MeetingService;
 import com.flodiback.global.rsData.RsData;
 
 import lombok.RequiredArgsConstructor;
@@ -22,22 +21,19 @@ public class MeetingController {
 
     @PostMapping
     public ResponseEntity<RsData<CreateMeetingResponse>> createMeeting(@RequestBody CreateMeetingRequest req) {
-        MeetingResult result = service.create(req.toCommand());
-        CreateMeetingResponse response = CreateMeetingResponse.from(result);
+        CreateMeetingResponse response = service.create(req);
         return ResponseEntity.status(HttpStatus.CREATED).body(RsData.of("201-1", "회의가 생성되었습니다.", response));
     }
 
     @PutMapping("/{id}/end")
     public ResponseEntity<RsData<MeetingDetailResponse>> endMeeting(@PathVariable Long id) {
-        MeetingResult result = service.end(id);
-        MeetingDetailResponse response = MeetingDetailResponse.from(result);
+        MeetingDetailResponse response = service.end(id);
         return ResponseEntity.ok(RsData.of("200-1", "회의가 종료되었습니다.", response));
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<RsData<MeetingDetailResponse>> getMeeting(@PathVariable Long id) {
-        MeetingResult result = service.getById(id);
-        MeetingDetailResponse response = MeetingDetailResponse.from(result);
+        MeetingDetailResponse response = service.getById(id);
         return ResponseEntity.ok(RsData.of("200-1", "회의 조회 성공.", response));
     }
 }
