--- src/main/java/com/flodiback/api/meeting/dto/CreateMeetingRequest.java
@@ -1,10 +1,0 @@
-package com.flodiback.api.meeting.dto;
-
-import com.flodiback.application.meeting.command.CreateMeetingCommand;
-
-public record CreateMeetingRequest(Long projectId, String title) {
-
-    public CreateMeetingCommand toCommand() {
-        return new CreateMeetingCommand(projectId, title);
-    }
-}
