+++ src/main/java/com/flodiback/domain/meeting/meeting/dto/MeetingDetailResponse.java
@@ -1,8 +1,8 @@
-package com.flodiback.application.meeting.result;
+package com.flodiback.domain.meeting.meeting.dto;
 
 import java.time.LocalDateTime;
 
-import com.flodiback.domain.meeting.type.MeetingStatus;
+import com.flodiback.global.enums.MeetingStatus;
 
-public record MeetingResult(
+public record MeetingDetailResponse(
         Long id, Long projectId, String title, LocalDateTime startedAt, LocalDateTime endedAt, MeetingStatus status) {}
