+++ src/test/java/com/flodiback/architecture/ArchitectureRulesTest.java
@@ -3,7 +3,6 @@ package com.flodiback.architecture;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.tngtech.archunit.core.importer.ImportOption;
@@ -34,23 +33,9 @@ class ArchitectureRulesTest {
                     "com.flodiback.global.rsData..",
                     "com.flodiback.global.globalExceptionHandler..",
                     "com.flodiback.api..",
-                    "com.flodiback.application..",
-                    "com.flodiback.infrastructure..",
+                    "org.springframework.http..",
                     "org.springframework.web..")
-            .because("도메인은 HTTP/API/응답 포맷과 기술 구현을 모르는 비즈니스 중심 계층이어야 합니다.");
-
-    @ArchTest
-    static final ArchRule application_must_not_depend_on_api_or_infrastructure = noClasses()
-            .that()
-            .resideInAPackage("com.flodiback.application..")
-            .should()
-            .dependOnClassesThat()
-            .resideInAnyPackage(
-                    "com.flodiback.api..",
-                    "com.flodiback.infrastructure..",
-                    "com.flodiback.global.rsData..",
-                    "org.springframework.web..")
-            .because("application 계층은 유스케이스를 조율하되 HTTP 계약과 기술 구현에 의존하지 않습니다.");
+            .because("도메인 기능 모듈은 HTTP/API/응답 포맷을 모르는 구현 영역으로 유지합니다.");
 
     @ArchTest
     static final ArchRule global_core_must_not_depend_on_project_layers = noClasses()
@@ -61,11 +46,7 @@ class ArchitectureRulesTest {
                     "com.flodiback.global.globalExceptionHandler..")
             .should()
             .dependOnClassesThat()
-            .resideInAnyPackage(
-                    "com.flodiback.api..",
-                    "com.flodiback.application..",
-                    "com.flodiback.domain..",
-                    "com.flodiback.infrastructure..")
+            .resideInAnyPackage("com.flodiback.api..", "com.flodiback.domain..")
             .because("공통 계층은 프로젝트 세부 계층에 의존하면 안 됩니다.");
 
     @ArchTest
@@ -75,12 +56,4 @@ class ArchitectureRulesTest {
             .should()
             .resideInAPackage("com.flodiback.domain..entity..")
             .because("@Entity 클래스는 탐색성과 일관성을 위해 entity 패키지에 둡니다.");
-
-    @ArchTest
-    static final ArchRule jpa_repositories_should_stay_in_infrastructure = classes()
-            .that()
-            .areAssignableTo(JpaRepository.class)
-            .should()
-            .resideInAPackage("com.flodiback.infrastructure.persistence..")
-            .because("Spring Data JPA 구현은 infrastructure persistence 계층에 둡니다.");
 }
